NEW Problem
When accessing different apps (readOnly and regular) from the sidebar,
it only bases whether to hide/show the "create a class" button on the
initial readOnly state of the first app clicked. e.g. when I switch from
TestApp to TestApp2 via the sidebar, the "create a class" button is
still there. If I enter TestApp2 first, the create a class button
doesn't appear for testApp if I switch via the sidebar